+++ jest.config.cjs
@@ -1,0 +1,19 @@
+module.exports = {
+    // testMatch: [],
+    collectCoverageFrom: [
+        'src/**/*.{mjs,js,jsx,ts,tsx}',
+        '!**/*.d.ts'
+    ],
+    setupFiles: [
+        '<rootDir>/test/bootstrap.cjs'
+    ],
+    testEnvironmentOptions: {
+        url: 'http://localhost:8080',
+    },
+    moduleNameMapper: {
+        '^(.*)\.js$': '$1',
+    },
+    transform: {
+        '\\.[jt]sx?$': ['babel-jest', { configFile: './test/babel.config.cjs' }]
+    },
+};