+++ src/dirname.mjs
@@ -1,6 +1,0 @@
-import { dirname, join } from "path";
-import { fileURLToPath } from "url";
-
-const __dirname = dirname(fileURLToPath(import.meta.url));
-
-export default __dirname;