+++ src/index.spec.ts
@@ -1,0 +1,7 @@
+import must from "must";
+
+describe("git-ahead", () => {
+    it("works", async () => {
+        must(true).be.true();
+    });
+});