+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
+The format is based on [EZEZ Changelog](https://ezez.dev/changelog/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED]
 (nothing yet)
+
+## [2.0.0] - 2025-01-06
+### Changed
+- output is grouped by project now
+- failing git-fetch won't stop the script from checking unpushed, uncommited and untracked changes
+### Dev
+- rewritten in TS
+- some preparations for future features
 
 ## [1.0.0] - 2020-06-10
 ### Added